refactor: move type to workflow-iterator
