+++ packages/twenty-front/src/modules/workflow/workflow-diagram/workflow-edges/utils/getConnectionOptionsForSourceHandle.ts
@@ -1,4 +1,4 @@
-import { type WorkflowStepConnectionOptions } from '@/workflow/workflow-diagram/types/WorkflowStepConnectionOptions';
+import { type WorkflowStepConnectionOptions } from '@/workflow/workflow-diagram/workflow-iterator/types/WorkflowStepConnectionOptions';
 import { WORKFLOW_DIAGRAM_ITERATOR_NODE_LOOP_HANDLE_ID } from '@/workflow/workflow-diagram/workflow-nodes/constants/WorkflowDiagramIteratorNodeLoopHandleId';
 
 export const getConnectionOptionsForSourceHandle = ({
