+++ packages/twenty-front/src/modules/workflow/workflow-diagram/workflow-iterator/types/WorkflowStepConnectionOptions.ts
